refactor: change product card
