+++ react-components/src/components/Card/Card.module.scss
@@ -1,25 +1,60 @@
-.layout {
+.card {
   display: flex;
   flex-direction: column;
-  height: 100%;
   width: 100%;
+  border-radius: 10px;
+  padding: 5px 10px;
+  gap: 10px;
+
+  background: white;
+  color: #1e1c33;
 }
 
-.content {
+.image {
+  width: 100%;
+  max-height: 100px;
+  object-fit: contain;
+}
+
+.title {
+  text-align: center;
+  font-weight: 500;
+  height: 48px;
+
+  white-space: wrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+
+}
+
+.description {
+  font-size: 15px;
+  padding-top: 10px;
+
+  white-space: wrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+
+}
+
+.cardFooter {
   display: flex;
-  flex-direction: row;
-  width: 100%;
+  justify-content: space-between;
+  align-items: center;
 }
 
-.childrenBlock {
-  max-width: 1200px;
-  margin: 0 auto;
-  width: 100%;
-  padding: 15px 30px;
+.rating {
+  display: flex;
+  gap: 10px;
+  font-weight: 500;
 }
 
-@media (max-width: 768px) {
-  .content {
-    display: block;
-  }
+.price {
+  font-size: 20px;
 }
