+++ react-components/src/pages/CardsPage/CardsPage.module.scss
@@ -1,0 +1,8 @@
+.cardWrapper {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 20px;
+    align-items: stretch;
+    width: 100%;
+  
+}
